Refactored cliTools
Moved cli-related functionality into a separate file.
